Create assignment 4.py
For this assignment you'll take what you've learned so far and build a classifier for bean disease. You'll be provided with training and validation data based on images taken of bean plants in Uganda. These images show healthy bean leaves as well as 2 types of common disease. Your job will be to build a neural network that can tell the difference between the healthy and diseased leaves.